make socket work with multiple locations
